Made the gamespeed framerate inependent.

+++ src/main/kotlin/game/spaceBalls/PowerUpSpawner.kt
@@ -35,11 +35,11 @@ class PowerUpSpawner(private val spaceBalls: SpaceBalls) {
                 RandomGenerator.randomInt(System.currentTimeMillis(),0, dropTable.size).also { index ->
                     println(index)
                     when(dropTable[index]){
-                        IPowerUp.PowerUpType.MED_KIT -> spaceBalls.powerUps.add(MedKit(randX, randY))
-                        IPowerUp.PowerUpType.SHIELD -> spaceBalls.powerUps.add(Shield(randX, randY))
-                        IPowerUp.PowerUpType.INVERTER -> spaceBalls.powerUps.add(Inverter(randX, randY))
+                        IPowerUp.PowerUpType.MED_KIT -> spaceBalls.powerUps.add(MedKit(randX.toDouble(), randY.toDouble()))
+                        IPowerUp.PowerUpType.SHIELD -> spaceBalls.powerUps.add(Shield(randX.toDouble(), randY.toDouble()))
+                        IPowerUp.PowerUpType.INVERTER -> spaceBalls.powerUps.add(Inverter(randX.toDouble(), randY.toDouble()))
                         IPowerUp.PowerUpType.CONTROL_INVERTER ->
-                            spaceBalls.powerUps.add(ControlInverter(randX, randY, spaceBalls))
+                            spaceBalls.powerUps.add(ControlInverter(randX.toDouble(), randY.toDouble(), spaceBalls))
                     }
                 }
             }
