+++ src/main/kotlin/game/spaceBalls/gameobjects/powerups/Inverter.kt
@@ -3,7 +3,7 @@ package main.kotlin.game.spaceBalls.gameobjects.powerups
 import main.kotlin.game.spaceBalls.gameobjects.GameObject
 import main.kotlin.game.spaceBalls.gameobjects.Player
 
-class Inverter(override val xPosition: Int, override val yPosition: Int) : GameObject, IPowerUp {
+class Inverter(override val xPosition: Double, override val yPosition: Double) : GameObject, IPowerUp {
 
     override fun tick() {
