+++ src/main/kotlin/game/spaceBalls/gameobjects/powerups/Shield.kt
@@ -3,7 +3,7 @@ package main.kotlin.game.spaceBalls.gameobjects.powerups
 import main.kotlin.game.spaceBalls.gameobjects.GameObject
 import main.kotlin.game.spaceBalls.gameobjects.Player
 
-class Shield(override val xPosition: Int, override val yPosition: Int) : GameObject, IPowerUp {
+class Shield(override val xPosition: Double, override val yPosition: Double) : GameObject, IPowerUp {
 
     companion object {
         const val AFFECTION_TIME = 10000
